Move concurrency test to own class
